+++ src/components/CardList/CardList.css
@@ -9,8 +9,27 @@
   padding: 0;
 }
 
+.card-list__button {
+  border: 0;
+  width: 320px;
+  height: 36px;
+  background-color: var(--color-secondary-bg);
+  border-radius: 6px;
+  color: var(--color-primary-txt);
+  font-weight: 500;
+  font-size: 12px;
+  line-height: 15px;
+  margin: 60px auto 80px;
+  cursor: pointer;
+}
+
+.card-list__button:hover {
+  opacity: 0.8;
+}
+
 @media (max-width: 1174px) {
   .cards-list {
     max-width: 730px;
   }
 } 
+
