+++ src/components/Preloader/Preloader.css
@@ -61,7 +61,7 @@
     position: absolute;
     top: 0;
     left: 0;
-    background: #3DDC84;
+    background: var(--color-primary);
     border-radius: 50px;
 }
