+++ src/components/SearchError/SearchError.css
@@ -1,6 +1,6 @@
 .search-error {
   margin: 0 auto;
-  min-height: 300px;
+  min-height: 150px;
 }
 
 .search-error__text {
